Use casts for accessing anonymous unions
Reuse the same 'cname' for all the structs and proxy their member with
dot-accesses. If one want to have a specific type in the union, it is
sufficient to cast to the desired type.

Make all event structs inherit from 'Event' so that they can reuse its
defined methods and attributes.